+++ MVbR2TG/MVbR2TG/libs/utilities/utilities.h
@@ -5,4 +5,4 @@
 #include <stdlib.h>
 
 void showImage (int binaryimage[IMG_HEIGHT][IMG_WIDTH]);
-void writeImage (int binaryimage[IMG_HEIGHT][IMG_WIDTH], FILE** _fptr, char const* filepath);
+void writeImage (short binaryimage[IMG_HEIGHT][IMG_WIDTH], FILE** _fptr, char const* filepath);
